move sqlite connection wrapper to new database.core module
