+++ src/maestral/database/orm.py
@@ -9,10 +9,10 @@ memory is constrained.
 
 from __future__ import annotations
 
-import sqlite3
 from weakref import WeakValueDictionary
 from typing import Any, Generator, TypeVar, Generic, Union, Type
 
+from .core import Database
 from .query import Query
 from .types import SqlType, SqlEnum, ColumnValueType
 
@@ -25,7 +25,6 @@ M = TypeVar("M", bound="Model")
 __all__ = [
     "Column",
     "NoDefault",
-    "Database",
     "Manager",
     "Model",
 ]
@@ -155,59 +154,6 @@ class Column(property):
         :returns: Converted Python value.
         """
         return self.type.sql_to_py(value)
-
-
-class Database:
-    """Proxy class to access sqlite3.connect method."""
-
-    def __init__(self, *args, **kwargs) -> None:
-        self.args = args
-        self.kwargs = kwargs
-        self._connection: sqlite3.Connection | None = None
-        self.Model = type(f"Model{self}", (Model,), {"_db": self})
-
-    @property
-    def connection(self) -> sqlite3.Connection:
-        """Returns an existing SQL connection or creates a new one."""
-
-        if self._connection:
-            return self._connection
-        else:
-            connection = sqlite3.connect(*self.args, **self.kwargs)
-            connection.row_factory = sqlite3.Row
-            self._connection = connection
-            return connection
-
-    def close(self) -> None:
-        """Closes the SQL connection."""
-        if self._connection:
-            self._connection.close()
-        self._connection = None
-
-    def commit(self) -> None:
-        """Commits SQL changes."""
-        self.connection.commit()
-
-    def execute(self, sql: str, *args) -> sqlite3.Cursor:
-        """
-        Creates a cursor and executes the given SQL statement.
-
-        :param sql: SQL statement to execute.
-        :param args: Parameters to substitute for placeholders in SQL statement.
-        :returns: The created cursor.
-        """
-        with self.connection:
-            return self.connection.execute(sql, args)
-
-    def executescript(self, script: str) -> None:
-        """
-        Creates a cursor and executes the given SQL script.
-
-        :param script: SQL script to execute.
-        :returns: The created cursor.
-        """
-        with self.connection:
-            self.connection.cursor().executescript(script)
 
 
 class Manager(Generic[M]):
